# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pharmaceutical RAG Chat API",
        "description": "\n\uc758\uc57d\ud488 RAG \ucc57\ubd07 \ubc31\uc5d4\ub4dc API\uc785\ub2c8\ub2e4.\n\n## \uc778\uc99d\n- \ucfe0\ud0a4 \uae30\ubc18 \uc138\uc158(`session` \ucfe0\ud0a4, HTTPOnly).\n- `POST /auth/register` \ub610\ub294 `POST /auth/login` \uc131\uacf5 \uc2dc \uc790\ub3d9\uc73c\ub85c \ucfe0\ud0a4\uac00 \uc124\uc815\ub429\ub2c8\ub2e4.\n- \ubaa8\ub4e0 \ubcf4\ud638\ub41c \uc5d4\ub4dc\ud3ec\uc778\ud2b8\ub294 \uc774 \ucfe0\ud0a4\ub97c \uc790\ub3d9 \ud3ec\ud568\ud569\ub2c8\ub2e4(`credentials: \"include\"`).\n\n## \uad8c\ud55c\n- `user` \u2014 \uc77c\ubc18 \uc0ac\uc6a9\uc790: \ucc44\ud305, \ubcf8\uc778 \ud504\ub85c\ud544 \uc870\ud68c/\uc218\uc815.\n- `admin` \u2014 \uad00\ub9ac\uc790: \uc704 \ubaa8\ub4e0 \uad8c\ud55c + \ud504\ub86c\ud504\ud2b8 \uc0dd\uc131/\ubc84\uc804 \uad00\ub9ac/\ud65c\uc131\ud654.\n\n## \ucc44\ud305 \uc2a4\ud2b8\ub9ac\ubc0d\n- `POST /chat/sessions/{session_id}/messages`\ub294 `text/event-stream` (SSE)\uc73c\ub85c \uc751\ub2f5\ud569\ub2c8\ub2e4.\n- \uc774\ubca4\ud2b8 \uc885\ub958: `token`, `tool`, `done`, `error`.\n",
        "version": "1.0.0"
    },
    "paths": {
...
        "/prompts": {
            "get": {
                "tags": [
                    "prompts"
                ],
                "summary": "\ud504\ub86c\ud504\ud2b8 \ubaa9\ub85d",
                "description": "\ub4f1\ub85d\ub41c \ubaa8\ub4e0 \ud504\ub86c\ud504\ud2b8\uc758 \uba54\ud0c0\ub370\uc774\ud130\ub97c \ubc18\ud658\ud569\ub2c8\ub2e4 (versions \ud544\ub4dc\ub294 \ube44\uc5b4\uc788\uc74c \u2014 \uc0c1\uc138\ub294 `GET /prompts/{key}`).",
                "operationId": "list_prompts_prompts_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/PromptPublic"
                                    },
                                    "type": "array",
                                    "title": "Response List Prompts Prompts Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "prompts"
                ],
                "summary": "\ud504\ub86c\ud504\ud2b8 \uc0dd\uc131 (admin)",
                "description": "\uc0c8 \ud504\ub86c\ud504\ud2b8 \ud0a4\ub97c \ub4f1\ub85d\ud569\ub2c8\ub2e4. \ubcf8\ubb38 \ubc84\uc804\uc740 `POST /prompts/{key}/versions`\ub85c \ucd94\uac00\ud558\uc138\uc694.",
                "operationId": "create_prompt_prompts_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PromptCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptPublic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "\uad8c\ud55c\uc774 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "\ub9ac\uc18c\uc2a4 \ucda9\ub3cc (\uc911\ubcf5 \ub4f1)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/prompts/{key}": {
            "get": {
                "tags": [
                    "prompts"
                ],
                "summary": "\ud504\ub86c\ud504\ud2b8 \uc0c1\uc138",
                "description": "\ud2b9\uc815 \ud504\ub86c\ud504\ud2b8\uc758 \ubaa8\ub4e0 \ubc84\uc804\uc744 \ud3ec\ud568\ud55c \uc0c1\uc138 \uc815\ubcf4\ub97c \ubc18\ud658\ud569\ub2c8\ub2e4.",
                "operationId": "get_prompt_prompts__key__get",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptPublic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "\ub9ac\uc18c\uc2a4\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/prompts/{key}/versions": {
            "post": {
                "tags": [
                    "prompts"
                ],
                "summary": "\ud504\ub86c\ud504\ud2b8 \ubc84\uc804 \uc0dd\uc131 (admin)",
                "description": "\uc0c8 \ubc84\uc804\uc744 \ucd94\uac00\ud569\ub2c8\ub2e4. `version_number`\ub294 \uc790\ub3d9\uc73c\ub85c \ub2e4\uc74c \ubc88\ud638\uac00 \ud560\ub2f9\ub418\uba70, \uae30\ubcf8 \ube44\ud65c\uc131 \uc0c1\ud0dc\uc785\ub2c8\ub2e4.",
                "operationId": "create_version_prompts__key__versions_post",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VersionCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VersionPublic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "\uad8c\ud55c\uc774 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "\ub9ac\uc18c\uc2a4\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/prompts/{key}/versions/{n}": {
            "delete": {
                "tags": [
                    "prompts"
                ],
                "summary": "\ud504\ub86c\ud504\ud2b8 \ubc84\uc804 \uc0ad\uc81c (admin)",
                "description": "\ud2b9\uc815 \ubc84\uc804\uc744 \uc0ad\uc81c\ud569\ub2c8\ub2e4. \ud65c\uc131 \ubc84\uc804\uc740 \uc0ad\uc81c\ud560 \uc218 \uc5c6\uc73c\ubbc0\ub85c, \uba3c\uc800 \ub2e4\ub978 \ubc84\uc804\uc744 \ud65c\uc131\ud654\ud574\uc57c \ud569\ub2c8\ub2e4.",
                "operationId": "delete_version_prompts__key__versions__n__delete",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Key"
                        }
                    },
                    {
                        "name": "n",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "N"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "\uad8c\ud55c\uc774 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "\ub9ac\uc18c\uc2a4\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\uc798\ubabb\ub41c \uc694\uccad \ubcf8\ubb38 \ub610\ub294 \ud30c\ub77c\ubbf8\ud130",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/prompts/{key}/activations": {
            "get": {
                "tags": [
                    "prompts"
                ],
                "summary": "\ud65c\uc131\ud654 \uc774\ub825 \uc870\ud68c (admin)",
                "description": "\uc774 \ud504\ub86c\ud504\ud2b8\uc758 \ud65c\uc131/\ube44\ud65c\uc131 \uc804\ud658 \uc774\ub825\uc744 \ucd5c\uc2e0\uc21c\uc73c\ub85c \ubc18\ud658\ud569\ub2c8\ub2e4.",
                "operationId": "list_activations_prompts__key__activations_get",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ActivationPublic"
                                    },
                                    "title": "Response List Activations Prompts  Key  Activations Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc558\uac70\ub098 \uc138\uc158\uc774 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "\uad8c\ud55c\uc774 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "\ub9ac\uc18c\uc2a4\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail",
                        "description": "\uc0ac\ub78c\uc774 \uc77d\uc744 \uc218 \uc788\ub294 \uc5d0\ub7ec \uba54\uc2dc\uc9c0",
                        "examples": [
                            "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ErrorResponse",
                "description": "\ubaa8\ub4e0 4xx/5xx \uc751\ub2f5\uc5d0\uc11c \uc0ac\uc6a9\ud558\ub294 \uacf5\ud1b5 \uc5d0\ub7ec \uc2a4\ud0a4\ub9c8."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PromptCreate": {
                "properties": {
                    "key": {
                        "type": "string",
                        "title": "Key",
                        "description": "\ud504\ub86c\ud504\ud2b8 \uc2dd\ubcc4 \ud0a4 (\uace0\uc720)",
                        "examples": [
                            "system_main"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "\ud504\ub86c\ud504\ud2b8 \uc6a9\ub3c4 \uc124\uba85",
                        "default": ""
                    }
                },
                "type": "object",
                "required": [
                    "key"
                ],
                "title": "PromptCreate"
            },
            "PromptPublic": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "key": {
                        "type": "string",
                        "title": "Key"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "versions": {
                        "items": {
                            "$ref": "#/components/schemas/VersionPublic"
                        },
                        "type": "array",
                        "title": "Versions",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "key",
                    "description",
                    "created_at"
                ],
                "title": "PromptPublic"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VersionCreate": {
                "properties": {
                    "content": {
                        "type": "string",
                        "title": "Content",
                        "description": "\ud504\ub86c\ud504\ud2b8 \ubcf8\ubb38"
                    },
                    "model": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model",
                        "description": "\uc774 \ubc84\uc804\uc5d0\uc11c \uc0ac\uc6a9\ud560 \ubaa8\ub378 ID. \ubbf8\uc9c0\uc815 \uc2dc \uc11c\ube44\uc2a4 \uae30\ubcf8\uac12",
                        "examples": [
                            "gpt-4o-mini"
                        ]
                    },
                    "temperature": {
                        "anyOf": [
                            {
                                "type": "number",
                                "maximum": 2.0,
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Temperature",
                        "description": "\uc0d8\ud50c\ub9c1 \uc628\ub3c4 (0.0 ~ 2.0). \ubbf8\uc9c0\uc815 \uc2dc \ubaa8\ub378 \uae30\ubcf8\uac12"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "VersionCreate"
            },
            "VersionPublic": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "version_number": {
                        "type": "integer",
                        "title": "Version Number"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "model": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model"
                    },
                    "temperature": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Temperature"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_by": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Created By"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "version_number",
                    "content",
                    "model",
                    "temperature",
                    "is_active",
                    "created_by",
                    "created_at"
                ],
                "title": "VersionPublic"
            }
        }
    },
...
}